+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in
 this file.
 
+## 1.0c - 2022/10/29
+### Added
+* added a function that refreshes the Twitch token upon expiry
+### Changed
+* changed authentication method on Twitch from an Implicit
+Grant Flow to Authorization Code Grant Flow
+
 ## 1.0b - 2022/10/25
 ### Fixed
 * forgot to include `user_settings.js` in the HTML file